+++ src/stdune/csexp.mli
@@ -10,5 +10,9 @@ val parse_string : string -> (t, string) Result.t
     canonical form to buffer [buf]. *)
 val to_buffer : buf:Buffer.t -> t -> unit
 
+(** [to_channel oc sexp] outputs the S-expression [sexp] converted to its
+    canonical form to channel [oc]. *)
+val to_channel : out_channel -> t -> unit
+
 (** [parse stream] reads one S-expression in canonical form from [stream] *)
 val parse : char Stream.t -> (t, string) Result.t